result-service/requirements.txt
flask
flask-cors
psycopg2-binary
gunicorn
pytest
redis
